move and rename AnimationProp file
