+++ packages/x-components/src/types/AnimationProp.ts
