+++ packages/x-components/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './origin';
 export * from './url-params';
+export * from './AnimationProp';
